+++ spec/trizetto/api/eligibility/web_service/uhc/patient_is_subscriber_spec.rb
@@ -1,5 +1,6 @@
-RSpec.describe "Tufts - Patient is the subscriber", type: :eligibility_response do
-  let(:file) {"tufts-health-plan/patient_is_subscriber.xml"}
+RSpec.describe "UHC - Patient is the subscriber", type: :eligibility_response do
+  let(:file) {"uhc/patient_is_subscriber.xml"}
 
+  let(:trace_number) {"851370272"}
   include_examples "active coverage"
 end